Extract `WebVitalStatusLineChart`
