+++ static/app/views/performance/browser/webVitals/webVitalsDetailPanel.tsx
@@ -1,12 +1,8 @@
 import {useMemo} from 'react';
 import {Link} from 'react-router';
-import {useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 
-import ChartZoom from 'sentry/components/charts/chartZoom';
-import MarkArea from 'sentry/components/charts/components/markArea';
-import MarkLine from 'sentry/components/charts/components/markLine';
-import {LineChart, LineChartSeries} from 'sentry/components/charts/lineChart';
+import {LineChartSeries} from 'sentry/components/charts/lineChart';
 import GridEditable, {
   COL_WIDTH_UNDEFINED,
   GridColumnHeader,
@@ -22,16 +18,11 @@ import {
 } from 'sentry/utils/performance/contexts/pageError';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
-import useRouter from 'sentry/utils/useRouter';
 import {PerformanceBadge} from 'sentry/views/performance/browser/webVitals/components/performanceBadge';
 import {WebVitalDescription} from 'sentry/views/performance/browser/webVitals/components/webVitalDescription';
+import {WebVitalStatusLineChart} from 'sentry/views/performance/browser/webVitals/components/webVitalStatusLineChart';
 import {calculateOpportunity} from 'sentry/views/performance/browser/webVitals/utils/calculateOpportunity';
-import {
-  calculatePerformanceScore,
-  PERFORMANCE_SCORE_MEDIANS,
-  PERFORMANCE_SCORE_P90S,
-} from 'sentry/views/performance/browser/webVitals/utils/calculatePerformanceScore';
+import {calculatePerformanceScore} from 'sentry/views/performance/browser/webVitals/utils/calculatePerformanceScore';
 import {
   Row,
   RowWithScore,
@@ -65,10 +56,6 @@ export function WebVitalsDetailPanel({
 }) {
   const organization = useOrganization();
   const location = useLocation();
-  const theme = useTheme();
-  const pageFilters = usePageFilters();
-  const router = useRouter();
-  const {period, start, end, utc} = pageFilters.selection.datetime;
   const transaction = location.query.transaction
     ? Array.isArray(location.query.transaction)
       ? location.query.transaction[0]
@@ -113,187 +100,16 @@ export function WebVitalsDetailPanel({
   const {data: timeseriesData, isLoading: isTimeseriesLoading} =
     useProjectWebVitalsValuesTimeseriesQuery({transaction});
 
-  const webVitalData: LineChartSeries[] = [
-    {
-      data:
-        !isTimeseriesLoading && webVital
-          ? timeseriesData?.[webVital].map(({name, value}) => ({
-              name,
-              value,
-            }))
-          : [],
-      seriesName: webVital ?? '',
-    },
-  ];
-
-  const showPoorMarkLine = webVitalData[0].data?.some(
-    ({value}) => value > PERFORMANCE_SCORE_MEDIANS[webVital ?? '']
-  );
-  const showMehMarkLine = webVitalData[0].data?.some(
-    ({value}) => value >= PERFORMANCE_SCORE_P90S[webVital ?? '']
-  );
-  const showGoodMarkLine = webVitalData[0].data?.every(
-    ({value}) => value < PERFORMANCE_SCORE_P90S[webVital ?? '']
-  );
-  const goodMarkArea = MarkArea({
-    silent: true,
-    itemStyle: {
-      color: theme.green300,
-      opacity: 0.1,
-    },
-    data: [
-      [
-        {
-          yAxis: PERFORMANCE_SCORE_P90S[webVital ?? ''],
-        },
-        {
-          yAxis: 0,
-        },
-      ],
-    ],
-  });
-  const mehMarkArea = MarkArea({
-    silent: true,
-    itemStyle: {
-      color: theme.yellow300,
-      opacity: 0.1,
-    },
-    data: [
-      [
-        {
-          yAxis: PERFORMANCE_SCORE_MEDIANS[webVital ?? ''],
-        },
-        {
-          yAxis: PERFORMANCE_SCORE_P90S[webVital ?? ''],
-        },
-      ],
-    ],
-  });
-  const poorMarkArea = MarkArea({
-    silent: true,
-    itemStyle: {
-      color: theme.red300,
-      opacity: 0.1,
-    },
-    data: [
-      [
-        {
-          yAxis: PERFORMANCE_SCORE_MEDIANS[webVital ?? ''],
-        },
-        {
-          yAxis: Infinity,
-        },
-      ],
-    ],
-  });
-  const goodMarkLine = MarkLine({
-    silent: true,
-    lineStyle: {
-      color: theme.green300,
-    },
-    label: {
-      formatter: () => 'Good',
-      position: 'insideEndBottom',
-      color: theme.green300,
-    },
-    data: showGoodMarkLine
-      ? [
-          [
-            {xAxis: 'min', y: 10},
-            {xAxis: 'max', y: 10},
-          ],
-        ]
-      : [
-          {
-            yAxis: PERFORMANCE_SCORE_P90S[webVital ?? ''],
-          },
-        ],
-  });
-  const mehMarkLine = MarkLine({
-    silent: true,
-    lineStyle: {
-      color: theme.yellow300,
-    },
-    label: {
-      formatter: () => 'Meh',
-      position: 'insideEndBottom',
-      color: theme.yellow300,
-    },
+  const webVitalData: LineChartSeries = {
     data:
-      showMehMarkLine && !showPoorMarkLine
-        ? [
-            [
-              {xAxis: 'min', y: 10},
-              {xAxis: 'max', y: 10},
-            ],
-          ]
-        : [
-            {
-              yAxis: PERFORMANCE_SCORE_MEDIANS[webVital ?? ''],
-            },
-          ],
-  });
-  const poorMarkLine = MarkLine({
-    silent: true,
-    lineStyle: {
-      color: theme.red300,
-    },
-    label: {
-      formatter: () => 'Poor',
-      position: 'insideEndBottom',
-      color: theme.red300,
-    },
-    data: [
-      [
-        {xAxis: 'min', y: 10},
-        {xAxis: 'max', y: 10},
-      ],
-    ],
-  });
-
-  webVitalData.push({
-    seriesName: '',
-    type: 'line',
-    markArea: goodMarkArea,
-    data: [],
-  });
-
-  webVitalData.push({
-    seriesName: '',
-    type: 'line',
-    markArea: mehMarkArea,
-    data: [],
-  });
-
-  webVitalData.push({
-    seriesName: '',
-    type: 'line',
-    markArea: poorMarkArea,
-    data: [],
-  });
-
-  webVitalData.push({
-    seriesName: '',
-    type: 'line',
-    markLine: goodMarkLine,
-    data: [],
-  });
-
-  webVitalData.push({
-    seriesName: '',
-    type: 'line',
-    markLine: mehMarkLine,
-    data: [],
-  });
-
-  if (showPoorMarkLine) {
-    webVitalData.push({
-      seriesName: '',
-      type: 'line',
-      markLine: poorMarkLine,
-      data: [],
-    });
-  }
+      !isTimeseriesLoading && webVital
+        ? timeseriesData?.[webVital].map(({name, value}) => ({
+            name,
+            value,
+          }))
+        : [],
+    seriesName: webVital ?? '',
+  };
 
   const detailKey = webVital;
 
@@ -395,32 +211,7 @@ export function WebVitalsDetailPanel({
           />
         )}
         <ChartContainer>
-          {webVital && (
-            <ChartZoom router={router} period={period} start={start} end={end} utc={utc}>
-              {zoomRenderProps => (
-                <LineChart
-                  {...zoomRenderProps}
-                  height={240}
-                  series={webVitalData}
-                  xAxis={{show: false}}
-                  grid={{
-                    left: 0,
-                    right: 15,
-                    top: 10,
-                    bottom: 0,
-                  }}
-                  yAxis={
-                    webVital === 'cls'
-                      ? {}
-                      : {axisLabel: {formatter: getFormattedDuration}}
-                  }
-                  tooltip={
-                    webVital === 'cls' ? {} : {valueFormatter: getFormattedDuration}
-                  }
-                />
-              )}
-            </ChartZoom>
-          )}
+          {webVital && <WebVitalStatusLineChart webVitalSeries={webVitalData} />}
         </ChartContainer>
         {!transaction && (
           <GridEditable
